Move specialized PEM helper into only extension that uses it
